deleted polidrome package
