+++ lib/main.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:presentation/my_app.dart';
 
-
-
-
 void main() {
   runApp(const MyApp());
 }
