+++ packages/domain/lib/usecase/login/usecase.dart
@@ -1,0 +1,5 @@
+abstract class UseCase<Params, Output> {
+  Output call(Params params);
+
+  void dispose();
+}
