+++ packages/domain/pubspec.yaml
@@ -1,0 +1,15 @@
+name: domain
+description: domain layer
+publish_to: none
+
+environment:
+  sdk: ">=2.16.1 <3.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+  rxdart: ^0.27.3
+
+
+  data:
+    path: ../data
