--- packages/plazo_mx/domain/pubspec.yaml
@@ -1,12 +1,0 @@
-name: domain
-description: domain layer
-publish_to: none
-
-environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=2.2.0"
-
-dependencies:
-  flutter:
-    sdk: flutter
-  rxdart: ^0.27.3
