--- packages/plazo_mx/presentation/pubspec.yaml
@@ -1,17 +1,0 @@
-name: presentation
-description: presentation layer
-publish_to: none
-
-environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=2.2.0"
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-  domain:
-    path: ../domain
-
-dev_dependencies: 
-  flutter_lints: ^1.0.0
