--- packages/polindrome/domain/lib/repository/base.dart
@@ -1,3 +1,0 @@
-abstract class Polindrome{
-  bool checkPalindrome(String inputString);
-}
