--- packages/polindrome/domain/lib/usecases/base_impl.dart
@@ -1,20 +1,0 @@
-
-import 'usecase.dart';
-
-class PalindromeUseCase implements UseCase<String, Future<bool>> {
-   @override
-   Future<bool> call(String params) async {
-     await Future.delayed(Duration(seconds: 5));
-     final isPalindrome = params == params.split('').reversed.toList().join('');
-     return isPalindrome;
-
-     //! Future.value(isPolindrome)
-   }
-
-   @override
-   void dispose() {
-   }
- }
-
-
-
