--- packages/polindrome/domain/lib/usecases/usecase.dart
@@ -1,5 +1,0 @@
-
-abstract class UseCase<Params, Output> {
-  Output call(Params params);
-  void dispose();
-}
