--- packages/polindrome/presentation/lib/pages/my_app.dart
@@ -1,24 +1,0 @@
-
-import 'package:flutter/material.dart';
-
-
-import 'home.dart';
-
-
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
-    );
-  }
-}
-
-
-
