--- packages/polindrome/presentation/lib/state_man/data/home_data.dart
@@ -1,7 +1,0 @@
-class HomeData {
-   bool? isPalindromeOrNot;
-   String palindromeInput = '';
-   HomeData(this.isPalindromeOrNot, this.palindromeInput);
-   factory HomeData.init() => HomeData(null, 'unselected');
-   HomeData copy() => HomeData(isPalindromeOrNot, palindromeInput);
- }
