--- packages/polindrome/presentation/pubspec.yaml
@@ -1,13 +1,0 @@
-name: presentation
-description: Presentation layer
-publish_to: none
-
-environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=2.2.0"
-
-dependencies:
-  flutter:
-    sdk: flutter
-  domain:
-    path: ../domain
