+++ packages/presentation/pubspec.yaml
@@ -1,0 +1,18 @@
+name: presentation
+description: presentation layer
+publish_to: none
+
+environment:
+  sdk: ">=2.16.1 <3.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+
+  domain:
+    path: ../domain
+  data:
+    path: ../data
+
+dev_dependencies: 
+  flutter_lints: ^1.0.0
