+++ pubspec.yaml
@@ -1,11 +1,9 @@
-name: polidrome
+name: mx_test_project
 description: A new Flutter project.
-
 
 publish_to: 'none' 
 
 version: 1.0.0+1
-
 environment:
   sdk: ">=2.16.1 <3.0.0"
 
@@ -14,13 +12,10 @@ dependencies:
   flutter:
     sdk: flutter
 
-
-
   cupertino_icons: ^1.0.2
-
   #local packages
   presentation:
-    path: packages/plazo_mx/presentation
+    path: packages/presentation
 
 
 dev_dependencies:
